rename usePhysics to useInitialConditions
