+++ src/hooks/index.ts
@@ -1,2 +1,2 @@
-export { usePhysics } from "./use-physics"
+export { useInitialConditions } from "./use-initial-conditions"
 export { useEntropy } from "./use-entropy"
